#CKFK20230409 Actualización de Gradle

+++ app/src/main/java/com/dts/roadp/clsFinDia.java
@@ -612,7 +612,7 @@ public class clsFinDia extends PBase{
 
         Cursor DT;
         boolean vFinDia = false;
-        int fechaUltimoCierre = ultimoCierreFecha();
+        long fechaUltimoCierre = ultimoCierreFecha();
 
         try{
